refactor: move back persistence_logic.dart
