+++ lib/services/link_service.dart
@@ -2,7 +2,7 @@ import 'package:flutter/services.dart';
 import 'package:lotti/database/database.dart';
 import 'package:lotti/features/tags/repository/tags_repository.dart';
 import 'package:lotti/get_it.dart';
-import 'package:lotti/logic/persistence/persistence_logic.dart';
+import 'package:lotti/logic/persistence_logic.dart';
 import 'package:lotti/services/tags_service.dart';
 
 class LinkService {
