+++ test/widgets/create/measurement_suggestions_test.dart
@@ -3,7 +3,7 @@ import 'package:lotti/classes/entity_definitions.dart';
 import 'package:lotti/classes/journal_entities.dart';
 import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
-import 'package:lotti/logic/persistence/persistence_logic.dart';
+import 'package:lotti/logic/persistence_logic.dart';
 import 'package:lotti/widgets/create/suggest_measurement.dart';
 import 'package:mocktail/mocktail.dart';
